added missing endpoints

+++ api/category/service_test.go
@@ -67,8 +67,8 @@ func TestService_GetCategories(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
@@ -148,8 +148,8 @@ func TestService_GetCategory(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
@@ -219,8 +219,8 @@ func TestService_GetCategoryForMonth(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
@@ -286,8 +286,8 @@ func TestService_GetCategoryForCurrentMonth(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
@@ -366,8 +366,8 @@ func TestService_UpdateCategoryForMonth(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
@@ -445,8 +445,8 @@ func TestService_UpdateCategoryForCurrentMonth(t *testing.T) {
 	assert.NoError(t, err)
 
 	var (
-		expectedGoalTarget             int64  = 18740
-		expectedGoalPercentageComplete uint16 = 20
+		expectedGoalTarget             int64 = 18740
+		expectedGoalPercentageComplete int32 = 20
 	)
 	expectedGoalCreationMonth, err := api.DateFromString("2018-04-01")
 	assert.NoError(t, err)
